Adding fingerprinting algorithm for adaptive throttling.
If PerConnBWFingerprint is non-zero, guard-to-client connections are
flagged if they use more than the FairEWMA portion of the bandwidth,
the EWMA a connection would have if it was using exactly its
fair share of bandwidth over time. Flagged connections are throttled at
the PerConnBWFingerprint rate, until their EWMA falls below
(PerConnBWPenalty * FairEWMA).

PerConnBWFingerprint and PerConnBWPenalty are config options, and
disabled by default.

This algorithm requires each connection to track EWMA. The EWMA halflife
for these connections can be set with PerConnHalflife and is used the
same way as CircuitPriorityHalflife. PerConnHalflife defaults to 30 if
not set and PerConnBWFingerprint is enabled.

PerConnHalflife cell counters can be tracked over time by setting config
PerConnHalflifeVerbose to non-zero. This results in a log entry EVERY
time the token buckets are refilled, so use with caution.