Prevent Libevent busy loop on write callback
A connection that is marked for close, wants to flush, is held open
to flush, but is rate limited (the token bucket is empty) will cause
busy loops in Libevent and infinite loops in the Shadow simulator. We
fix this by de-registering the write event with Libevent and marking the
connection as blocked on bandwidth.