Prevent Libevent busy loop on read callback
A connection that is marked for close, wants to flush, is held open
to flush, but is rate limited (the token bucket is empty) will cause
busy loops in Libevent and infinite loops in the Shadow simulator. We
fix this by de-registering the read event with Libevent and marking the
connection as blocked on bandwidth.

See also commit 287dfaa8f3b535f81b3b3f0c2b7079ee7015eaa2