Add thread pool for running viterbi on streams
When counting traffic models with PrivCount, we can now use a
thread pool to run the viterbi algorithm on completed streams.
This offloads the expensive viterbi computation off of the main
thread. We use Tor's workqueue and threadpool implemention to
set up the threads and manage the jobs.

A new config option called 'PrivCountNumViterbiWorkers' was added
and must be set to a positive value in order to enable worker
threads that can run viterbi. Because of the way Tor's thread pool
works, the number of threads cannot be changed after the first
time it is initialized. So, PrivCountNumViterbiWorkers should be
set appropriately when starting the Tor process.

Changes to PrivCountNumViterbiWorkers after the process starts
are handled gracefully in the viterbi code. However, the size of
the thread pool will not change if it was previously initialized.
So, going from a positive value to a different positive value
has no effect until the process is restarted. Going from a
positive value to 0 causes the code to ignore the thread pool
and run the viterbi jobs in the main thread. Going back to a
positive value will again use the initial number of threads to
run the jobs.