bugfix on memleak during PrivCount tmodel measurement
We change the location of the tmodel_stream_free() call to ensure
that it gets called for all streams for which the object exists.
This removes the memleak that was consuming GiBs of RAM over hours
of running the relay.

We also changed the location of tmodel_stream_new() to that tmodel
only functions on non-hs streams.

Updated the tmodel test case to use TT_FORK rather than running in
the main thread of the test runner.

Finally, we no longer override 'F' on streams that only have a
single packet. This is to deal with traffic models that do not
define any 'F' observations in their starting states.