Support for testing Tor relay speeds
This commit implements a new relay speedtest feature that runs on
a Tor client to test the speed of a Tor relay. The intended use
case is that several Tor clients and relays will run this branch
and "team up" to test the speed of other Tor relays.

The client and the last relay in a speedtest circuit must support the
speedtest by running this branch. The other relays in a speedtest
circuit can run other versions of Tor.

The following new control commands are used to control a speedtest:
SPEEDTEST OPEN fp1,fp2,...
SPEEDTEST START circid lensecs
SPEEDTEST STOP circid
SPEEDTEST CLOSE circid

Where fpi is the 40 character fingerprint (without the $) of relay i,
and lensecs is the number of seconds to run the test.

Again, the last relay in the OPEN command must be running this branch.
The intended use-case is to run a client and relay on the same machine,
and then build a two hop circuit that ends at the test relay. The
test will isolate the speed of the first hop in the circuit.

Once a speed test is started, the client sends speedtest cells until
the circuit is blocked on streams. Whenever the circuit becomes
unblocked, the client resumes sending speedtest cells. The client and
relay also track the number of speedtest cells processed and logs them
to a file. The client also emits this information via a asynchronous
control events.

The new async control events that can be tracked are:
650 SPEEDTEST OPENED circid fp,fp,...
650 SPEEDTEST STARTED circid startts lensecs stopts fp,fp,...
650 SPEEDTEST STOPPED circid startts lensecs stopts cellsrecv cellssent fp,fp,...
650 SPEEDTEST CLOSED circid cellsrecvtot cellssenttot fp,fp,...

Other notes:

- We prevent scheduler from overflowing conn outbuf

This is a severe bug that in the best case can cause significant
increase in memory usage on an edge, and in the worst case can
cause channels to get destroyed if the outbuf is overflowed.

- The speedtest clients do not encrypt speedtest cells

Stopping the client from encrypting speedtest cells means the
relay will decrypt and then will not recognize the cell. This will
stop the relay from doing CPU intensive tasks associated with
recognized cells, like additional parsing and digest checks. The
relay will simply forward back the unrecognized cell as before.

- Client speedtest circuits timeout 10 minutes after opening